+++ core/src/globals.rs
@@ -1,0 +1,1 @@
+pub const INVINVILIBITY_ALPHA_RANGE: std::ops::RangeInclusive<f32> = 0.25..=1.0;
