+++ core/src/prelude.rs
@@ -1,8 +1,8 @@
 pub use {
     crate::{
-        attachment::*, bullet::*, camera::*, damage::*, debug::*, debug::*, elements::*, input::*,
-        item::*, item::*, lifetime::*, map::*, math::*, metadata::*, physics::*, player::*,
-        session::*, MAX_PLAYERS,
+        attachment::*, bullet::*, camera::*, damage::*, debug::*, debug::*, elements::*,
+        globals::*, input::*, item::*, item::*, lifetime::*, map::*, metadata::*, physics::*,
+        player::*, session::*, utils::*, MAX_PLAYERS,
     },
     bones_bevy_asset::{BevyAssets, BonesBevyAsset, BonesBevyAssetLoad},
     bones_lib::prelude::*,
